+++ server/src/shared/models/product.model.ts
@@ -5,6 +5,7 @@ import {
 } from '@shared/models/base.model';
 import { CategoryModel } from '@shared/models/category.model';
 import { MediaModel } from '@shared/models/media.model';
+import { OptionModel } from '@shared/models/option.model';
 import { SupplierModel } from '@shared/models/supplier.model';
 import { z } from 'zod';
 
@@ -23,6 +24,7 @@ export const ProductModel = MetadataFields.extend({
       resource: MediaModel,
     }),
   ),
+  options: z.array(OptionModel).optional(),
   category: CategoryModel.optional(),
   supplier: SupplierModel,
 });
