+++ web/src/app/admin/product/[id]/page.tsx
@@ -1,8 +1,12 @@
-import { ReactNode } from 'react';
-type ProductDetailAdminPageProps = {};
+type ProductDetailAdminPageProps = {
+  params: Promise<{ id: string }>;
+};
 
-const ProductDetailAdminPage = ({}: ProductDetailAdminPageProps) => {
-  return <div></div>;
+const ProductDetailAdminPage = async ({
+  params,
+}: ProductDetailAdminPageProps) => {
+  const { id } = await params;
+  return <div>{id}</div>;
 };
 
 export default ProductDetailAdminPage;
